Add ability to set switch value in manifest cliFlags property
main.js adds command line switches to the spawned electron process if "cliFlags" property has values in electron.manifest.json, but it does not allow setting a value for the switch. This change splits each item in the cliFlags array on the first '=' and sets the switch value, if provided.